+++ changelog.md
@@ -2,4 +2,4 @@
 - Fixed crash on dedicated server
 
 ### Wireless Chargers 1.0.9
-- Initial release of Wireless Chargers for Minecraft 1.21.5
+- Initial release of Wireless Chargers for Minecraft 1.21.6